# packages/contracts/gen/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "platform aggregate API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "x-request-body-limit": 16384,
        "operationId": "createUserProfile",
        "summary": "provision a profile row for an existing sub",
        "x-required-permission": "user:create:any",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Profile"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "profile already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listUsers",
        "summary": "paginated profiles with validated server-side sorting",
        "x-required-permission": "user:read:any",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "enum": [
                10,
                20,
                50
              ],
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "case-insensitive email/display-name search",
            "schema": {
              "type": "string",
              "maxLength": 120
            }
          },
          {
            "name": "presence",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "online",
                "offline"
              ]
            }
          },
          {
            "name": "roleId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 1
            }
          },
          {
            "name": "registeredFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "registeredTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "displayName",
                "email",
                "lastLoginAt"
              ],
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "items": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Profile"
                              }
                            },
                            "meta": {
                              "$ref": "#/components/schemas/EnvelopeMeta"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "getUser",
        "x-required-permission": "user:read:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Profile"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "patch": {
        "x-request-body-limit": 16384,
        "operationId": "updateUser",
        "x-auth": "required",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/EnvelopeOK"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Profile"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "soft-delete the profile row and emit user.deleted",
        "x-required-permission": "user:delete:any",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeOK"
                }
              }
            }
          },
          "404": {
            "description": "unknown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeFail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "EnvelopeMeta": {
        "type": "object",
        "required": [
          "limit",
          "offset",
          "total"
        ],
        "properties": {
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "EnvelopeOK": {
        "type": "object",
        "required": [
          "success",
          "message",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "data": {
            "description": "endpoint-specific payload"
          }
        }
      },
      "EnvelopeFail": {
        "type": "object",
        "required": [
          "success",
          "message",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "maxLength": 255
          },
          "error": {
            "type": "string",
            "maxLength": 120
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "maxLength": 120
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 400,
            "pattern": "^https://"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "deleted"
            ]
          },
          "lockedUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleSummary"
            }
          },
          "online": {
            "type": "boolean"
          },
          "activeSessions": {
            "type": "integer"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastLoginIp": {
            "type": "string",
            "maxLength": 45
          },
          "lastLoginUserAgent": {
            "type": "string",
            "maxLength": 512
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RoleSummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "maxLength": 60
          }
        }
      },
...
      "ProfileInput": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "existing subject id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "maxLength": 120
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "maxLength": 400,
            "pattern": "^https://"
          }
        }
      },
...
    }
  }
}